Fix circular dep compilation error
